onlinePredictionTest/requirements.txt
tensorflow==1.0.1
google-api-python-client==1.5.1
oauth2client==2.2.0
google-cloud-pubsub==0.25.0
